Model Controllers refactor done, no build errors, testing
