+++ RVNav/AddVehicle/Storyboards/AddVehicleViewController.swift
@@ -33,7 +33,7 @@ class AddVehicleViewController: ShiftableViewController {
     var vehicle: Vehicle?
     var vehicles: [Vehicle]?
     var avoidance: [Avoid] = []
-    var networkController: NetworkControllerProtocol = WebRESTAPINetworkController()
+    var modelController = ModelController()
     
     // MARK: - View LifeCycle
     
@@ -131,7 +131,7 @@ class AddVehicleViewController: ShiftableViewController {
         
         let newVehicle = Vehicle(id: nil, name: vehicleName, height: vehicleHeight, weight: vehicleWeight, width: vehicleWidth, length: vehicleLength, axelCount: vehicleAxelCount, vehicleClass: vehicleType, dualTires: duelWheelSwitch.isOn, trailer: nil)
         
-        networkController.createVehicle(with: newVehicle) { error in
+        modelController.vehicleController.createVehicle(with: newVehicle) { error in
                             if let error = error {
                 NSLog("Error Creating Vehicle \(error)")
             }
