+++ RVNav/ModelControllers/AvoidanceControllerProtocol.swift
@@ -1,0 +1,14 @@
+//
+//  AvoidanceControllerProtocol.swift
+//  RVNav
+//
+//  Created by Lambda_School_Loaner_214 on 1/16/20.
+//  Copyright © 2020 RVNav. All rights reserved.
+//
+
+import Foundation
+
+
+protocol AvoidanceControllerProtocol {
+    func getAvoidances(with routeInfo: RouteInfo, completion: @escaping ([Avoid]?,Error?) -> Void)
+}
