+++ RVNav/ModelControllers/UserControllerProtocol.swift
@@ -9,6 +9,8 @@
 import Foundation
 
 protocol UserControllerProtocol {
+    var result: Result? {get set}
+    
     func register(with user: User, completion: @escaping (Error?) -> Void)
     func signIn(with signInInfo: SignInInfo, completion: @escaping (Error?) -> Void)
     func logout(completion: @escaping () -> Void)
