+++ RVNav/Signup+SignIn/View Controllers/Onboarding2ViewController.swift
@@ -13,7 +13,7 @@ class Onboarding2ViewController: ShiftableViewController {
     
     // MARK: - Properties
     var formData: FormData?
-    var networkController: NetworkControllerProtocol!
+    var userController: UserControllerProtocol!
     
     // MARK: - IBOutlets
     @IBOutlet weak var backgroundImageView: UIImageView!
@@ -50,7 +50,7 @@ class Onboarding2ViewController: ShiftableViewController {
                         password: formData.password,
                         email: formData.email,
                         username: formData.username)
-        networkController?.register(with: user) { (error) in
+        userController?.register(with: user) { (error) in
             DispatchQueue.main.async {
                 if let error = error {
                     NSLog("\(error)")
@@ -65,7 +65,7 @@ class Onboarding2ViewController: ShiftableViewController {
                         !password.isEmpty else { return }
                     
                     let signInInfo = SignInInfo(email: email, password: password)
-                    self.networkController.signIn(with: signInInfo) { (error) in
+                    self.userController.signIn(with: signInInfo) { (error) in
                         if let error = error {
                             NSLog("Error signing in: \(error)")
                         } else {
